// src/services/proxy/generated/models/Autor.ts
/* tslint:disable */
/* eslint-disable */
/**
 * The Golden Book API
 * API para negocio de venta de libros 
 *
 * The version of the OpenAPI document: 1.0
 * Contact: soporte@thegoldenbook.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface Autor
 */
export interface Autor {
    /**
     * 
     * @type {number}
     * @memberof Autor
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof Autor
     */
    nombre?: string;
    /**
     * 
     * @type {string}
     * @memberof Autor
     */
    apellido1?: string;
    /**
     * 
     * @type {string}
     * @memberof Autor
     */
    apellido2?: string;
    /**
     * 
     * @type {Date}
     * @memberof Autor
     */
    fechaNacimiento?: Date;
}

/**
 * Check if a given object implements the Autor interface.
 */
export function instanceOfAutor(value: object): value is Autor {
    return true;
}

export function AutorFromJSON(json: any): Autor {
    return AutorFromJSONTyped(json, false);
}

export function AutorFromJSONTyped(json: any, ignoreDiscriminator: boolean): Autor {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'nombre': json['nombre'] == null ? undefined : json['nombre'],
        'apellido1': json['apellido1'] == null ? undefined : json['apellido1'],
        'apellido2': json['apellido2'] == null ? undefined : json['apellido2'],
        'fechaNacimiento': json['fechaNacimiento'] == null ? undefined : (new Date(json['fechaNacimiento'])),
    };
}

export function AutorToJSON(json: any): Autor {
    return AutorToJSONTyped(json, false);
}

export function AutorToJSONTyped(value?: Autor | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'nombre': value['nombre'],
        'apellido1': value['apellido1'],
        'apellido2': value['apellido2'],
        'fechaNacimiento': value['fechaNacimiento'] == null ? undefined : ((value['fechaNacimiento']).toISOString()),
    };
}