// src/services/proxy/generated/models/Tematica.ts
/* tslint:disable */
/* eslint-disable */
/**
 * The Golden Book API
 * API para negocio de venta de libros 
 *
 * The version of the OpenAPI document: 1.0
 * Contact: soporte@thegoldenbook.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface Tematica
 */
export interface Tematica {
    /**
     * 
     * @type {number}
     * @memberof Tematica
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof Tematica
     */
    nombre?: string;
}

/**
 * Check if a given object implements the Tematica interface.
 */
export function instanceOfTematica(value: object): value is Tematica {
    return true;
}

export function TematicaFromJSON(json: any): Tematica {
    return TematicaFromJSONTyped(json, false);
}

export function TematicaFromJSONTyped(json: any, ignoreDiscriminator: boolean): Tematica {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'nombre': json['nombre'] == null ? undefined : json['nombre'],
    };
}

export function TematicaToJSON(json: any): Tematica {
    return TematicaToJSONTyped(json, false);
}

export function TematicaToJSONTyped(value?: Tematica | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'nombre': value['nombre'],
    };
}