// src/services/proxy/generated/models/ValoracionDTO.ts
export interface ValoracionDTO {
    /**
     * 
     * @type {number}
     * @memberof ValoracionDTO
     */
    clienteId?: number;
    /**
     * 
     * @type {string}
     * @memberof ValoracionDTO
     */
    nickname?: string;
    /**
     * 
     * @type {string}
     * @memberof ValoracionDTO
     */
    nombreCliente?: string;
    /**
     * 
     * @type {string}
     * @memberof ValoracionDTO
     */
    apellido1Cliente?: string;
    /**
     * 
     * @type {string}
     * @memberof ValoracionDTO
     */
    apellido2Cliente?: string;
    /**
     * 
     * @type {number}
     * @memberof ValoracionDTO
     */
    libroId?: number;
    /**
     * 
     * @type {string}
     * @memberof ValoracionDTO
     */
    nombreLibro?: string;
    /**
     * 
     * @type {number}
     * @memberof ValoracionDTO
     */
    autorId?: number;
    /**
     * 
     * @type {string}
     * @memberof ValoracionDTO
     */
    nombreAutor?: string;
    /**
     * 
     * @type {string}
     * @memberof ValoracionDTO
     */
    apellido1Autor?: string;
    /**
     * 
     * @type {string}
     * @memberof ValoracionDTO
     */
    apellido2Autor?: string;
    /**
     * 
     * @type {number}
     * @memberof ValoracionDTO
     */
    numeroEstrellas?: number;
    /**
     * 
     * @type {string}
     * @memberof ValoracionDTO
     */
    asunto?: string;
    /**
     * 
     * @type {string}
     * @memberof ValoracionDTO
     */
    cuerpo?: string;
    /**
     * 
     * @type {Date}
     * @memberof ValoracionDTO
     */
    fechaPublicacion?: Date;
}